♻️ MJKWG1 code: decompose incident resolver strategies
