+++ .agentplane/tasks/202604171502-MJKWG1/pr/github-title.txt
@@ -1,0 +1,1 @@
+refactor/runtime: Decompose runtime incidents resolver into strategy modules (MJKWG1)
