+++ .agentplane/tasks/202604171502-MJKWG1/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202604171502-MJKWG1/incident-strategies",
+  "created_at": "2026-04-17T18:44:54.616Z",
+  "head_sha": "b0c53c8aae1a7e985ed3734ab3a02380b62e0d0e",
+  "last_verified_at": "2026-04-17T18:53:09.885Z",
+  "last_verified_sha": "b0c53c8aae1a7e985ed3734ab3a02380b62e0d0e",
+  "schema_version": 1,
+  "task_id": "202604171502-MJKWG1",
+  "updated_at": "2026-04-17T18:44:54.616Z",
+  "verify": {
+    "status": "pass"
+  }
+}
