+++ .agentplane/tasks/202604171502-MJKWG1/pr/review.md
@@ -1,0 +1,58 @@
+# PR Review
+
+Created: 2026-04-17T18:44:54.616Z
+Branch: task/202604171502-MJKWG1/incident-strategies
+
+## Summary
+
+Decompose runtime incidents resolver into strategy modules
+
+Split the runtime incidents resolver hotspot into declarative strategy modules so the dispatcher stays small and behavior remains unchanged.
+
+## Scope
+
+- In scope: Split the runtime incidents resolver hotspot into declarative strategy modules so the dispatcher stays small and behavior remains unchanged.
+- Out of scope: unrelated refactors not required for "Decompose runtime incidents resolver into strategy modules".
+
+## Verification
+
+### Plan
+
+1. Run `bunx vitest run packages/agentplane/src/runtime/incidents/resolve.test.ts --hookTimeout 60000 --testTimeout 60000`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Run `bun run typecheck`. Expected: it succeeds and confirms the requested outcome for this task.
+3. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+4. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Decomposed runtime incidents resolver into registry, planning, and advice strategy modules; reduced resolve.ts to a façade re-export surface; verified typecheck plus resolve incident contract.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-04-17T18:44:54.616Z
+- Branch: task/202604171502-MJKWG1/incident-strategies
+- Head: b0c53c8aae1a
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
